% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.R
\name{bicluster}
\alias{bicluster}
\title{Bicluster data}
\usage{
bicluster(data, P0, Q0, miss_val, miss_val_sd = 1, row_min_num = 5,
  col_min_num = 5, row_num_to_move = 1, col_num_to_move = 1,
  row_shuffles = 1, col_shuffles = 1, max.iter = 100, verbose = TRUE)
}
\arguments{
\item{data}{Dataset to bicluster. Must to be a data matrix with only numbers and missing values in the data set. It should have row names and column names.}

\item{P0}{Initial matrix for column prototypes, maybe generated by `partition_gen`.}

\item{Q0}{Initial matrix for row prototypes, maybe generated by `partition_gen`.}

\item{miss_val}{Value or function to put in empty cells of the prototype matrix.
If a value, a random normal variable with sd = `miss_val_sd` is used each iteration.}

\item{miss_val_sd}{Standard deviation of the normal distribution `miss_val` follows
if `miss_val` is a number. By default this equals 1.}

\item{row_min_num}{Minimum row prototype size in order to be eligible to be chosen when filling an empty row prototype. Default is 5.}

\item{col_min_num}{Minimum column prototype size in order to be eligible to be chosen when filling an empty row prototype. Default is 5.}

\item{row_num_to_move}{Number of rows to remove from the sampled prototype to put in the empty row prototype. Default is 1.}

\item{col_num_to_move}{Number of columns to remove from the sampled prototype to put in the empty column prototype. Default is 1.}

\item{row_shuffles}{Number of times to shuffle rows in each iteration. Default is 1.}

\item{col_shuffles}{Number of times to shuffle columns in each iteration. Default is 1.}

\item{max.iter}{Maximum number of iterations to let the algorithm run for.}

\item{verbose}{Logical. If TRUE, will report progress.}
}
\value{
A list containing final matrices for column and
    row partitions, the SSE of the original partitioning, the SSE for each iteration, the Rand Indices for row and column
    prototypes, the number of iterations the algorithm ran for, and the final prototype matrix.
}
\description{
Bicluster data
}
\examples{
dat <- kronecker(matrix(1:6, nrow = 2, ncol = 3), matrix(5, nrow = 3, ncol = 4))
dat[sample(1:length(dat), 0.5 * length(dat))] <- NA
dat <- dat[sample(1:nrow(dat), nrow(dat)), sample(1:ncol(dat), ncol(dat))]
P01 <- partition_gen(12, 3)
Q01 <- partition_gen(6, 2)

bc <- bicluster(dat, P01, Q01, miss_val = mean(dat, na.rm = TRUE),
                miss_val_sd = sd(dat, na.rm = TRUE),
                col_min_num = 2, row_min_num = 2,
                col_num_to_move = 1, row_num_to_move = 1,
                max.iter = 10)
bc
gg_sse(bc)
gg_bicluster(bc, dat)
}
