% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclustermd-package.R
\docType{package}
\name{biclustermd-package}
\alias{biclustermd-package}
\title{biclustermd: A package to bicluster data with missing values}
\description{
The main function is \code{biclustermd()}. Results can be plotted with \code{autoplot()}
and \code{as.Biclust()} converts results to Biclust objects.
}
