% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_vector_gen.R
\name{binary_vector_gen}
\alias{binary_vector_gen}
\title{Make a binary vector with all values equal to zero except for one}
\usage{
binary_vector_gen(n, i)
}
\arguments{
\item{n}{Desired vector length.}

\item{i}{Index whose value is one.}
}
\value{
A vector
}
\description{
Make a binary vector with all values equal to zero except for one
}
