% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_iteration_sum_sse.R
\name{cluster_iteration_sum_sse}
\alias{cluster_iteration_sum_sse}
\title{Calculate the sum cluster SSE in each iteration}
\usage{
cluster_iteration_sum_sse(data, P, Q)
}
\arguments{
\item{data}{The data being biclustered. Must to be a data matrix with only numbers and missing values in the data set. It should have row names and column names.}

\item{P}{Matrix for column prototypes.}

\item{Q}{Matrix for row prototypes.}
}
\value{
The SSE for the parameters specified.
}
\description{
Calculate the sum cluster SSE in each iteration
}
