% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.names.biclustermd.R
\name{col.names.biclustermd}
\alias{col.names.biclustermd}
\title{Get data matrix column names and their corresponding column cluster membership}
\usage{
\method{col.names}{biclustermd}(x)
}
\arguments{
\item{x}{and object of class \code{biclustermd}}
}
\value{
a data frame with column names of the shuffled matrix and corresponding column cluster names.
}
\description{
Get data matrix column names and their corresponding column cluster membership
}
\examples{
data("synthetic")
# default parameters
bc <- biclustermd(synthetic)
bc
col.names(bc)
# this is a simplified version of the output for gather(bc):
library(dplyr)
gather(bc) \%>\% distinct(col_cluster, col_name)
}
