% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.names.R
\name{col.names}
\alias{col.names}
\title{A generic to gather column names}
\usage{
col.names(x)
}
\arguments{
\item{x}{an object to retrieve column names from}
}
\description{
A generic to gather column names
}
