% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_partition.R
\name{format_partition}
\alias{format_partition}
\title{Format a partition matrix}
\usage{
format_partition(P1)
}
\arguments{
\item{P1}{A partition matrix.}
}
\value{
A formatted partition matrix.
}
\description{
Formats a partition matrix so that subsets in a partition will be ordered by the value of the smallest in each subset
}
