% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRI.R
\name{gg_ri}
\alias{gg_ri}
\title{Plot each iteration's Rand Indices using the ggplot2 framework}
\usage{
gg_ri(x, ...)
}
\arguments{
\item{x}{A bicluster object.}

\item{...}{Arguments to be passed to `geom_point()` and `geom_line()`.}
}
\value{
A ggplot
}
\description{
Plot each iteration's Rand Indices using the ggplot2 framework
}
