% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot.biclustermd}
\alias{ggplot.biclustermd}
\title{Plot the SSE or Rand Indices of a \code{biclustermd} object.}
\usage{
\method{ggplot}{biclustermd}(data, mapping = NULL, value = c("sse",
  "similarity"), ..., environment = NULL)
}
\arguments{
\item{data}{an object of class \code{biclustermd}.}

\item{mapping}{unused; included to match \code{ggplot2} generic}

\item{value}{which value to plot. Can be either "sse" or "similarity".}

\item{...}{unused; included to match \code{ggplot2} generic}

\item{environment}{unused; included to match \code{ggplot2} generic}
}
\value{
A data frame
}
\description{
Plot the SSE or Rand Indices of a \code{biclustermd} object.
}
