% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Compute the Jaccard similarity coefficient for two clusterings}
\usage{
jaccard_similarity(clus1, clus2)
}
\arguments{
\item{clus1}{vector giving the first set of clusters}

\item{clus2}{vector giving the second set of clusters}
}
\value{
a numeric
}
\description{
Compute the Jaccard similarity coefficient for two clusterings
}
\references{
Milligan, G.W. and Cooper, M. C. (1986) \emph{A study of the comparability of external criteria for hierarchical cluster analysis.
Multivariate Behavioral Research, 21, 441-458.}
}
