% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_bicluster.R
\name{old_bicluster}
\alias{old_bicluster}
\title{Bicluster data (old)}
\usage{
old_bicluster(data, P0, Q0, miss_val, col_min_num = 5,
  row_min_num = 10, col_num_to_move = 5, row_num_to_move = 5,
  max.iter = 100, verbose = TRUE)
}
\arguments{
\item{data}{Dataset to bicluster. Must to be a data matrix with only numbers and missing values in the data set. It should have row names and column names.}

\item{P0}{Initial matrix for column prototypes, maybe generated by `partition_gen`.}

\item{Q0}{Initial matrix for row prototypes, maybe generated by `partition_gen`.}

\item{miss_val}{Value or function to put in empty cells of the prototype matrix.}

\item{col_min_num}{Minimum column prototype size in order to be eligible to be chosen when filling an empty row prototype.}

\item{row_min_num}{Minimum row prototype size in order to be eligible to be chosen when filling an empty row prototype. Default is 10.}

\item{col_num_to_move}{Number of columns to remove from the sampled prototype to put in the empty column prototype.}

\item{row_num_to_move}{Number of rows to remove from the sampled prototype to put in the empty row prototype. Default is 5.}

\item{max.iter}{Maximum number of iterations to let the algorithm run for.}

\item{verbose}{Logical. If TRUE, will report progress.}
}
\value{
A list containing final matrices for column and row prototypes, the SSE for each iteration, the Rand Indices for row and column prototypes, the number of iterations the algorithm ran for, and the final prototype matrix.
}
\description{
Bicluster data (old)
}
