% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_matrix_to_vector.R
\name{part_matrix_to_vector}
\alias{part_matrix_to_vector}
\title{Convert a partition matrix to a vector}
\usage{
part_matrix_to_vector(P0)
}
\arguments{
\item{P0}{A partition matrix}
}
\value{
An integer vector
}
\description{
For each row in a partition matrix, this function gets the column index for which the row is equal to one.
That is, for row i, this function returns the index of the row entry that is equal to one.
}
