% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_finder.R
\name{position_finder}
\alias{position_finder}
\title{Find the index of the first nonzero value in a vector}
\usage{
position_finder(vec)
}
\arguments{
\item{vec}{A binary vector}
}
\value{
Position of the first nonzero value in a vector.
}
\description{
Find the index of the first nonzero value in a vector
}
