% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.biclustermd}
\alias{print.biclustermd}
\title{Print an object of class biclustermd}
\usage{
\method{print}{biclustermd}(x, ...)
}
\arguments{
\item{x}{a \code{biclustermd} object.}

\item{...}{arguments passed to or from other methods}
}
\description{
Print an object of class biclustermd
}
