% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIPlots.R
\name{RI_plots}
\alias{RI_plots}
\title{Plot each iteration's Rand Indices}
\usage{
RI_plots(bicluster_obj, pch = 20, cex = 0.5, lwd = 1)
}
\arguments{
\item{bicluster_obj}{A bicluster object.}

\item{pch}{Point shape. This is equivalent to the `pch` argument for base graphics.}

\item{cex}{Scale of the points in the plot.}

\item{lwd}{Width of the smoothing spline line.}
}
\value{
A plot
}
\description{
Plot each iteration's Rand Indices
}
