% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row.names.biclustermd.R
\name{row.names.biclustermd}
\alias{row.names.biclustermd}
\title{Get data matrix row names and their corresponding row cluster membership}
\usage{
\method{row.names}{biclustermd}(x)
}
\arguments{
\item{x}{and object of class \code{biclustermd}}
}
\value{
a data frame with row names of the shuffled matrix and corresponding row cluster names.
}
\description{
Get data matrix row names and their corresponding row cluster membership
}
\examples{
data("synthetic")
# default parameters
bc <- biclustermd(synthetic)
bc
row.names(bc)
# this is a simplified version of the output for gather(bc):
library(dplyr)
gather(bc) \%>\% distinct(row_cluster, row_name)
}
