% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumSSEplots.R
\name{sumSSE_plots}
\alias{sumSSE_plots}
\title{Plot each iteration's sum of SSEs and a smoothing spline fit to the data.}
\usage{
sumSSE_plots(x, linear = FALSE, pch = 20, cex = 1, lwd = 3)
}
\arguments{
\item{x}{A bicluster object.}

\item{linear}{Logical. If `TRUE`, a linear fit is plotted.}

\item{pch}{Point shape. This is equivalent to the `pch` argument for base graphics.}

\item{cex}{Scale of the points in the plot.}

\item{lwd}{Width of lines.}
}
\value{
A plot
}
\description{
Plot each iteration's sum of SSEs and a smoothing spline fit to the data.
}
