% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\docType{data}
\name{synthetic}
\alias{synthetic}
\title{Synthetic data for examples.}
\format{
An object of class \code{matrix} with 6 rows and 12 columns.
}
\usage{
synthetic
}
\description{
This simple dataset allows users to use data that are easy to understand while learning
\code{biclustermd}. This is a matrix with 6 rows and 12 columns. 50\% of values are missing.
}
\keyword{datasets}
