% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellMSE.R
\name{cell_mse}
\alias{cell_mse}
\title{Make a data frame containing the MSE for each bicluster cell}
\usage{
cell_mse(x, data)
}
\arguments{
\item{x}{A bicluster object.}

\item{data}{Data the bicluster object was generated from}
}
\value{
A data frame
}
\description{
Make a data frame containing the MSE for each bicluster cell
}
