% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColNames.R
\name{col_cluster_names}
\alias{col_cluster_names}
\title{Get column names in each column cluster}
\usage{
col_cluster_names(bc_object, data)
}
\arguments{
\item{bc_object}{Biclustering object to extract column cluster designation from}

\item{data}{Data that contains the column names}
}
\value{
A data frame with two columns: \code{cluster} corresponds to the column
  cluster and \code{name} gives the column names in each cluster.
}
\description{
Get column names in each column cluster
}
