% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColNames.R
\name{col_cluster_names}
\alias{col_cluster_names}
\title{Get column names in each column cluster}
\usage{
col_cluster_names(bc_object, data)
}
\arguments{
\item{bc_object}{Biclustering object to extract column cluster designation from}

\item{data}{Data that contains the column names}
}
\value{
A data frame with two columns: \code{cluster} corresponds to the column
  cluster and \code{name} gives the column names in each cluster.
}
\description{
Get column names in each column cluster
}
\examples{
data("synthetic")
rownames(synthetic) <- letters[1:nrow(synthetic)]
colnames(synthetic) <- letters[1:col(synthetic)]
P01 <- partition_gen(12, 3)
Q01 <- partition_gen(6, 2)

bc <- bicluster(synthetic, P01, Q01, miss_val = mean(synthetic, na.rm = TRUE),
                miss_val_sd = sd(synthetic, na.rm = TRUE),
                col_min_num = 2, row_min_num = 2,
                col_num_to_move = 1, row_num_to_move = 1,
                max.iter = 10)
row_cluster_names(bc, synthetic)
col_cluster_names(bc, synthetic)
}
