% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather.biclustermd}
\alias{gather.biclustermd}
\title{Gather a biclustermd object}
\usage{
\method{gather}{biclustermd}(x)
}
\arguments{
\item{x}{a \code{biclustermd} object to gather.}
}
\value{
A data frame containing the row names and column names of both the 
  two-way table of data biclustered and the cell-average matrix.
}
\description{
Gather a biclustermd object
}
