% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbicluster.R
\name{gg_bicluster}
\alias{gg_bicluster}
\title{Make a heatmap of sparse biclustering results}
\usage{
gg_bicluster(bc_object, data, transform_colors = TRUE, c = 2/15)
}
\arguments{
\item{bc_object}{A bicluster object.}

\item{data}{The raw data that was biclustered.}

\item{transform_colors}{If equals `TRUE` (default) then the data is scaled by 
`c` and run through a standard normal cdf before plotting. If `FALSE`, raw data
values are used in the heat map.}

\item{c}{Value to scale the data by before running it through a standard normal CDF. Default is 2/15.}
}
\value{
An object of class ggplot.
}
\description{
Make a heatmap of sparse biclustering results
}
