% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSSE.R
\name{gg_sse}
\alias{gg_sse}
\title{Plot each iteration's SSE and a smoother or linear model fit to the data.}
\usage{
gg_sse(bicluster_obj, linear = FALSE, pt_size = 1.25, line_width = 1)
}
\arguments{
\item{bicluster_obj}{A bicluster object.}

\item{linear}{Logical. If TRUE, a line is added to the plot. By default it is FALSE and no line is added to the plot.}

\item{pt_size}{Point size. It's default value is 1.25.}

\item{line_width}{Width of lines. Default is 1.}
}
\value{
A ggplot.
}
\description{
Plot each iteration's SSE and a smoother or linear model fit to the data.
}
