% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSSE.R
\name{gg_sse}
\alias{gg_sse}
\title{Plot each iteration's SSE.}
\usage{
gg_sse(bicluster_obj, ..., smoother = FALSE, linear = FALSE)
}
\arguments{
\item{bicluster_obj}{A bicluster object.}

\item{...}{Arguments to be passed to `geom_point`}

\item{smoother}{Logical. If TRUE, a smoother is added to the plot.
By default it is FALSE and no smoother is added to the plot.}

\item{linear}{Logical. If TRUE, a line is added to the plot.
By default it is FALSE and no line is added to the plot.}
}
\value{
A ggplot.
}
\description{
Plot each iteration's SSE.
}
