% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTrainingSet.R
\name{make_train_set}
\alias{make_train_set}
\title{Make a data frame containing information about what cluster a row-column pair is in}
\usage{
make_train_set(data, bicluster_obj)
}
\arguments{
\item{data}{Data that was fed into the bicluster function.}

\item{bicluster_obj}{A bicluster object.}
}
\value{
A data frame
}
\description{
Make a data frame containing information about what cluster a row-column pair is in
}
