% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEHeatmap.R
\name{mse_heatmap}
\alias{mse_heatmap}
\title{Make a heatmap of cell MSEs}
\usage{
mse_heatmap(x, linewidth = 0.1, log_scale = FALSE)
}
\arguments{
\item{x}{An object of class \code{biclustermd}.}

\item{linewidth}{Width of vertical and horizontal lines. Default is 0.1.}

\item{log_scale}{Logical. If TRUE, fill legend will be on a log10 scale. Default is FALSE.}
}
\value{
A ggplot object.
}
\description{
Make a heatmap of cell MSEs
}
\examples{
data("synthetic")
P01 <- partition_gen(12, 3)
Q01 <- partition_gen(6, 2)

bc <- bicluster(synthetic, P01, Q01, miss_val = mean(synthetic, na.rm = TRUE),
                miss_val_sd = sd(synthetic, na.rm = TRUE),
                col_min_num = 2, row_min_num = 2,
                col_num_to_move = 1, row_num_to_move = 1,
                max.iter = 10)
mse_heatmap(bc)
}
