% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEHeatmap.R
\name{mse_heatmap}
\alias{mse_heatmap}
\title{Make a heatmap of cell MSEs}
\usage{
mse_heatmap(mse_obj, linewidth = 0.1, log10_col = TRUE)
}
\arguments{
\item{mse_obj}{A data frame returned by `cell_mse`.}

\item{linewidth}{Width of vertical and horizontal lines. Default is 0.1.}

\item{log10_col}{Logical. If TRUE (default), fill legend will be on a log10 scale.}
}
\value{
A ggplot object.
}
\description{
Make a heatmap of cell MSEs
}
