% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEHeatmap.R
\name{mse_heatmap}
\alias{mse_heatmap}
\title{Make a heatmap of cell MSEs}
\usage{
mse_heatmap(mse_obj)
}
\arguments{
\item{mse_obj}{A data frame returned by `cell_mse`}
}
\value{
A ggplot object.
}
\description{
Make a heatmap of cell MSEs
}
