% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionGenByP.R
\name{partition_gen_by_p}
\alias{partition_gen_by_p}
\title{Create a partition matrix with a partition vector p}
\usage{
partition_gen_by_p(N, K, p)
}
\arguments{
\item{N}{Rows in a partition matrix}

\item{K}{Number of prototypes to create}

\item{p}{Integer vector containing the prototype each row in a partition matrix is to be assigned to.}
}
\value{
A partition matrix.
}
\description{
Create a partition matrix with a partition vector p
}
