% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderBiclust.R
\name{reorder_biclust}
\alias{reorder_biclust}
\title{Reorder a bicluster object for making a heat map}
\usage{
reorder_biclust(bc_object)
}
\arguments{
\item{bc_object}{A bicluster object.}
}
\value{
A list containing the two partition matrices used by gg_bicluster.
}
\description{
Reorder a bicluster object for making a heat map
}
