% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIPlots.R
\name{RI_plots}
\alias{RI_plots}
\title{Plot each iteration's Rand Indices}
\usage{
RI_plots(x, ...)
}
\arguments{
\item{x}{A bicluster object.}

\item{...}{Arguments to be passed to base R line and point graphics.}
}
\value{
A plot
}
\description{
Plot each iteration's Rand Indices
}
