% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRowNames.R
\name{row_cluster_names}
\alias{row_cluster_names}
\title{Get row names in each row cluster}
\usage{
row_cluster_names(bc_object, data)
}
\arguments{
\item{bc_object}{Biclustering object to extract row cluster designation from}

\item{data}{Data that contains the row names}
}
\value{
A data frame with two columns: \code{cluster} corresponds to the row
  cluster and \code{name} gives the row names in each cluster.
}
\description{
Get row names in each row cluster
}
\examples{
dat <- data("synthetic")
rownames(dat) <- letters[1:nrow(dat)]
colnames(dat) <- letters[1:col(dat)]
P01 <- partition_gen(12, 3)
Q01 <- partition_gen(6, 2)

bc <- bicluster(dat, P01, Q01, miss_val = mean(dat, na.rm = TRUE),
                miss_val_sd = sd(dat, na.rm = TRUE),
                col_min_num = 2, row_min_num = 2,
                col_num_to_move = 1, row_num_to_move = 1,
                max.iter = 10)
row_cluster_names(bc, dat)
col_cluster_names(bc, dat)
}
