% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRowNames.R
\name{row_cluster_names}
\alias{row_cluster_names}
\title{Get row names in each row cluster}
\usage{
row_cluster_names(bc_object, data)
}
\arguments{
\item{bc_object}{Biclustering object to extract row cluster designation from}

\item{data}{Data that contains the row names}
}
\value{
A data frame with two columns: \code{cluster} corresponds to the row
  cluster and \code{name} gives the row names in each cluster.
}
\description{
Get row names in each row cluster
}
